+++ test/startNext.mjs
@@ -1,5 +1,6 @@
 import { createServer } from "http";
 import { createRequire } from "module";
+
 import listen from "./listen.mjs";
 
 const require = createRequire(import.meta.url);