+++ changelog.md
@@ -2,8 +2,14 @@
 
 ## Next
 
+### Major
+
+- Updated Node.js support to `^14.17.0 || ^16.0.0 || >= 18.0.0`.
+
 ### Patch
 
+- Updated GitHub Actions CI config:
+  - Run tests with Node.js v14, v16, v18.
 - Fixed a readme typo.
 
 ## 12.0.0