+++ changelog.md
@@ -10,6 +10,8 @@
 
 - Updated GitHub Actions CI config:
   - Run tests with Node.js v14, v16, v18.
+  - Updated `actions/checkout` to v3.
+  - Updated `actions/setup-node` to v3.
 - Fixed a readme typo.
 
 ## 12.0.0