+++ changelog.md
@@ -1,6 +1,6 @@
 # next-graphql-react changelog
 
-## Next
+## 13.0.0
 
 ### Major